% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.treatment.R
\name{continuous.treatment}
\alias{continuous.treatment}
\title{continuous treatment indicator function}
\usage{
continuous.treatment(input.level, output.level, coordinate.system = 4326,
  filter = NULL, filter.automatic = FALSE)
}
\arguments{
\item{input.level}{An sf data frame with geometries and associated binary treatment indicator. The treatment indicator column should be named "treatment".}

\item{output.level}{An sf data frame with geometries of the geographic units the function should output a continuous treatment indicator for.}

\item{coordinate.system}{Four digit epsg code of the projection you wish to use. Defaults to 4326.}

\item{filter}{Smallest area an intersection between the input.level and output.level geographic units may have to be considered valid. All intersections with a smaller area will be filtered out. Defaults to NULL.}

\item{filter.automatic}{If set to TRUE the function discards all intersections between input.level and output.level geographic units that have a smaller area than the smallest input.level geographic unit. Defaults to FALSE.}
}
\value{
A data frame of output.level geographic units and the percentage to which they were treated.
}
\description{
Given a binary treatment indicator for some set of geographic units, this function allows you to
aggregate the treatment data to a set of larger scale geographic units and create a continuous treatment
indicator based on the fraction of the larger geographic units' area that was treated.
Considering potential for mismatches in geometries between geographic units the function allows you
to filter out "invalid" overlaps between geographic units arising as an artefact of these mismatches,
by way of a user set filter value or a generic inbuilt method.
}
