% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.intersections.R
\name{percent.intersections}
\alias{percent.intersections}
\title{percent overlap function}
\usage{
percent.intersections(input.level, output.level,
  coordinate.system = 4326, filter = NULL, filter.automatic = FALSE)
}
\arguments{
\item{input.level}{An sf data frame.}

\item{output.level}{An sf data frame.}

\item{coordinate.system}{Four digit epsg code of the projection you wish to use. Defaults to 4326.}

\item{filter}{Smallest area an intersection between the input.level and output.level geographic units may have to be considered valid. All intersections with a smaller area will be filtered out. Defaults to NULL.}

\item{filter.automatic}{If set to TRUE the function discards all intersections between input.level and output.level geographic units that have a smaller area than the smallest input.level geographic unit. Defaults to FALSE.}
}
\value{
A data frame named "intersections" with a column "area_percent" indicating the percentage area of an output.level geographic unit covered by a specific input.level geographic unit.
}
\description{
This function allows you to compute the percentage area of a larger scale geographic 
unit that is covered by a smaller scale geographic unit.
Considering potential for mismatches in geometries between geographic units the function allows you
to filter out "invalid" overlaps between geographic units arising as an artefact of these mismatches,
by way of a user set filter value or a generic inbuilt method.
}
